requirements.txt
Django==4.1.13
djangorestframework==3.15.1
psycopg2==2.9.9
django-taggit==5.0.1
django-embed-video==1.4.10
Pillow==10.3.0
social-auth-app-django==5.4.1